move pkg cmd/helm/strvals to pkg/strvals
This is another useful package outside of cmd/helm.
