+++ cmd/helm/upgrade.go
@@ -25,10 +25,10 @@ import (
 	"github.com/ghodss/yaml"
 	"github.com/spf13/cobra"
 
-	"k8s.io/helm/cmd/helm/strvals"
 	"k8s.io/helm/pkg/chartutil"
 	"k8s.io/helm/pkg/helm"
 	"k8s.io/helm/pkg/storage/driver"
+	"k8s.io/helm/pkg/strvals"
 )
 
 const upgradeDesc = `
