+++ pkg/strvals/doc.go
